Move ownershipvoucher under services
