+++ pkg/services/ownershipvoucher/deserializer.go
